// FinalTask/types/JobDetails.ts
export interface JobDetails {
	id: string;
	title: string;
	description: string;
	responsibilities: string;
	requirements: string;
	idealCandidate: string;
	categories: string[];
	opType: string;
	startDate: string;
	endDate: string;
	deadline: string;
	location: string[];
	requiredSkills: string[];
	whenAndWhere: string;
	orgID: string;
	datePosted: string;
	status: string;
	applicantsCount: number;
	viewsCount: number;
	orgName: string;
	logoUrl: string;
	isBookmarked: boolean;
	isRolling: boolean;
	questions: null | string;
	perksAndBenefits: null | string;
	createdAt: string;
	updatedAt: string;
	orgPrimaryPhone: string;
	orgEmail: string;
	orgWebsite: string;
	average_rating: number;
	total_reviews: number;
}